requirements.txt
apache-airflow==2.3.2
tensorflow==2.8.0
pandas
numpy
sqlalchemy
psycopg2-binary
streamlit
scikit-learn
matplotlib
seaborn